+++ mds/apis/prv_api/providers.py
@@ -17,7 +17,15 @@ class ProviderSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = models.Provider
-        fields = ("id", "name", "logo_b64")
+        fields = (
+            "id",
+            "name",
+            "logo_b64",
+            "base_api_url",
+            "api_authentication",
+            "api_configuration",
+            "agency_api_authentication",
+        )
 
 
 class ProviderViewSet(viewsets.ModelViewSet):