+++ mds/factories.py
@@ -573,6 +573,9 @@ class Provider(factory.DjangoModelFactory):
         ["Lime", "BlueLA", "Metro Bike", "LongProviderNameCompanyLtdSarlGmoLgbtq"]
     )
     base_api_url = "http://provider"
+    api_configuration = {"bluela_dash": False, "trailing_slash": False}
+    api_authentication = {"type": "none"}
+    agency_api_authentication = {"type": "none"}
 
 
 class Device(factory.DjangoModelFactory):