+++ tests/apis/prv_api/test_providers.py
@@ -32,21 +32,32 @@ def test_provider_basic(client, django_assert_num_queries):
         response = client.get("/prv/providers/", **auth_header(SCOPE_ADMIN))
     assert response.status_code == 200
     assert len(response.data) == 3
-
     assert {
         "id": "aaaaaaa0-1342-413b-8e89-db802b2f83f6",
         "name": "Test provider",
         "logo_b64": None,
+        "base_api_url": "http://provider",
+        "api_configuration": {"bluela_dash": False, "trailing_slash": False},
+        "api_authentication": {"type": "none"},
+        "agency_api_authentication": {"type": "none"},
     } in response.data
     assert {
         "id": "bbbbbbb0-1342-413b-8e89-db802b2f83f6",
         "name": "Test provider 2",
         "logo_b64": None,
+        "base_api_url": "http://provider",
+        "api_configuration": {"bluela_dash": False, "trailing_slash": False},
+        "api_authentication": {"type": "none"},
+        "agency_api_authentication": {"type": "none"},
     } in response.data
     assert {
         "id": "ccccccc0-1342-413b-8e89-db802b2f83f6",
         "name": "Test provider 3",
         "logo_b64": None,
+        "base_api_url": "http://provider",
+        "api_configuration": {"bluela_dash": False, "trailing_slash": False},
+        "api_authentication": {"type": "none"},
+        "agency_api_authentication": {"type": "none"},
     } in response.data
 
     response = client.get(
@@ -58,4 +69,8 @@ def test_provider_basic(client, django_assert_num_queries):
         "id": "aaaaaaa0-1342-413b-8e89-db802b2f83f6",
         "name": "Test provider",
         "logo_b64": None,
+        "base_api_url": "http://provider",
+        "api_configuration": {"bluela_dash": False, "trailing_slash": False},
+        "api_authentication": {"type": "none"},
+        "agency_api_authentication": {"type": "none"},
     }